Changes detailed in the Proposed June changes have been made.
See actual Database Changes June 2019.pdf for details. Note that these changes includea new function ST_DISTANCE_SPHERE which is not needed if you are running a later version of MYSQL/MariaDB - the function will be dropped in future. meanwhile it is there to enable us to locate the nearest device to any given coordinates.
If starting with a fresh database just source this to create all the tables.